refactor(types): share easymotion target shape
